Move builtin plugin registry into separate header.
