+++ Engine/plugin/agsplugin.cpp
@@ -46,6 +46,7 @@
 #include "media/audio/sound.h"
 #include "plugin/agsplugin.h"
 #include "plugin/plugin_engine.h"
+#include "plugin/plugin_builtin.h"
 #include "plugin/pluginobjectreader.h"
 #include "script/script.h"
 #include "script/script_runtime.h"
