+++ Engine/plugin/plugin_engine.h
@@ -40,19 +40,6 @@ bool pl_is_plugin_loaded(const char *pl_name);
 //returns whether _any_ plugins want a particular event
 bool pl_any_want_hook(int event);
 
-//  Initial implementation for apps to register their own inbuilt plugins
-
-struct InbuiltPluginDetails {
-    char      filename[PLUGIN_FILENAME_MAX+1];
-    void      (*engineStartup) (IAGSEngine *);
-    void      (*engineShutdown) ();
-    int       (*onEvent) (int, int);
-    void      (*initGfxHook) (const char *driverName, void *data);
-    int       (*debugHook) (const char * whichscript, int lineNumber, int reserved);
-};
-
-// Register a builtin plugin.
-int pl_register_builtin_plugin(InbuiltPluginDetails const &details);
 void pl_set_file_handle(long data, AGS::Common::Stream *stream);
 void pl_clear_file_handle();
